Reorganized the code we wrote so far
