+++ CircuitMessing/app/build.gradle
@@ -49,6 +49,7 @@ dependencies {
     implementation 'androidx.lifecycle:lifecycle-extensions:2.2.0'
     implementation 'androidx.navigation:navigation-fragment-ktx:2.3.1'
     implementation 'androidx.navigation:navigation-ui-ktx:2.3.1'
+    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.1.1'
 
     // Import the Firebase BoM
     implementation platform('com.google.firebase:firebase-bom:25.12.0')
