+++ CircuitMessing/app/src/main/java/com/example/circuitmessing/products/makerbuino/MakerbuinoActivity.kt
@@ -3,7 +3,6 @@ package com.example.circuitmessing.products.makerbuino
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.view.MenuItem
 import android.view.View
 import android.widget.TextView
 import androidx.appcompat.widget.Toolbar
@@ -14,19 +13,12 @@ import androidx.fragment.app.FragmentTransaction
 import com.escaper.escaper.utils.preferences
 import com.example.circuitmessing.MainActivity
 import com.example.circuitmessing.R
-import com.example.circuitmessing.products.nibble.Nibble_time_to_get_makin
-import com.example.circuitmessing.products.nibble.Nibble_introduction_fragment
+import com.example.circuitmessing.products.ProgressManager.Companion.checkDonePages
+import com.example.circuitmessing.products.ProgressManager.Companion.updatePageDone
 import com.google.android.material.navigation.NavigationView
-import com.google.firebase.database.DataSnapshot
-import com.google.firebase.database.DatabaseError
-import com.google.firebase.database.DatabaseReference
-import com.google.firebase.database.ValueEventListener
-import com.google.firebase.database.ktx.database
-import com.google.firebase.ktx.Firebase
 import kotlinx.android.synthetic.main.activity_ringo.*
 
 class MakerbuinoActivity : AppCompatActivity() {
-    private lateinit var database: DatabaseReference
     private var mFragmentManager: FragmentManager = supportFragmentManager
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -34,7 +26,7 @@ class MakerbuinoActivity : AppCompatActivity() {
         setContentView(R.layout.activity_makerbuino)
         val toolbar: Toolbar = findViewById(R.id.toolbar)
         setSupportActionBar(toolbar)
-        database = Firebase.database.reference
+
 
         val drawerLayout: DrawerLayout = findViewById(R.id.drawer_layout)
         val navView: NavigationView = findViewById(R.id.nav_view)
@@ -53,13 +45,13 @@ class MakerbuinoActivity : AppCompatActivity() {
                     returnHome()
                 }
                 R.id.nav_introduction -> {
-                    replaceFragment(R.id.maker_buino_fragment, Makerbuino_introduction())
+                    replaceFragment(R.id.maker_buino_fragment, MakerbuinoIntroductionFragment())
                 }
                 R.id.nav_meet_tools -> {
-                    replaceFragment(R.id.maker_buino_fragment, Makerbuino_meet_the_tools())
+                    replaceFragment(R.id.maker_buino_fragment, MakerbuinoMeetTheToolsFragment())
                 }
                 R.id.nav_time_makin -> {
-                    replaceFragment(R.id.maker_buino_fragment, Makerbuino_time_to_get_makin())
+                    replaceFragment(R.id.maker_buino_fragment, MakerbuinoTimeToGetMakinFragment())
                 }
                 R.id.nav_summed_up -> {
                     replaceFragment(R.id.maker_buino_fragment, MakerbuinoFinishingUpFragment())
@@ -87,7 +79,7 @@ class MakerbuinoActivity : AppCompatActivity() {
         checkDonePages(productName = "makerbuino", pageName = "makin", item3)
         checkDonePages(productName = "makerbuino", pageName = "summed", item4)
 
-        replaceFragment(R.id.maker_buino_fragment, Makerbuino_introduction())
+        replaceFragment(R.id.maker_buino_fragment, MakerbuinoIntroductionFragment())
 
         // Update database when specific page is done
         rightArrow?.setOnClickListener {
@@ -95,13 +87,13 @@ class MakerbuinoActivity : AppCompatActivity() {
                 navView.menu.getItem(1).isChecked -> {
                     val item = navView.menu.getItem(1)
                     updatePageDone(productName = "makerbuino", pageName = "intro", item)
-                    replaceFragment(R.id.maker_buino_fragment, Makerbuino_meet_the_tools())
+                    replaceFragment(R.id.maker_buino_fragment, MakerbuinoMeetTheToolsFragment())
                     navView.menu.getItem(2).isChecked = true;
                 }
                 navView.menu.getItem(2).isChecked -> {
                     val item = navView.menu.getItem(2)
                     updatePageDone(productName = "makerbuino", pageName = "tools", item)
-                    replaceFragment(R.id.maker_buino_fragment, Makerbuino_time_to_get_makin())
+                    replaceFragment(R.id.maker_buino_fragment, MakerbuinoTimeToGetMakinFragment())
                     navView.menu.getItem(3).isChecked = true;
                 }
                 navView.menu.getItem(3).isChecked -> {
@@ -127,19 +119,19 @@ class MakerbuinoActivity : AppCompatActivity() {
                 navView.menu.getItem(2).isChecked -> {
                     val item = navView.menu.getItem(1)
                     updatePageDone(productName = "makerbuino", pageName = "tools", item)
-                    replaceFragment(R.id.maker_buino_fragment, Makerbuino_introduction())
+                    replaceFragment(R.id.maker_buino_fragment, MakerbuinoIntroductionFragment())
                     navView.menu.getItem(1).isChecked = true;
                 }
                 navView.menu.getItem(3).isChecked -> {
                     val item = navView.menu.getItem(2)
                     updatePageDone(productName = "makerbuino", pageName = "makin", item)
-                    replaceFragment(R.id.maker_buino_fragment, Makerbuino_meet_the_tools())
+                    replaceFragment(R.id.maker_buino_fragment, MakerbuinoMeetTheToolsFragment())
                     navView.menu.getItem(2).isChecked = true;
                 }
                 navView.menu.getItem(4).isChecked -> {
                     val item = navView.menu.getItem(3)
                     updatePageDone(productName = "makerbuino", pageName = "summed", item)
-                    replaceFragment(R.id.maker_buino_fragment, Makerbuino_time_to_get_makin())
+                    replaceFragment(R.id.maker_buino_fragment, MakerbuinoTimeToGetMakinFragment())
                     navView.menu.getItem(3).isChecked = true;
                 }
                 navView.menu.getItem(5).isChecked -> {
@@ -158,49 +150,6 @@ class MakerbuinoActivity : AppCompatActivity() {
         val intent = Intent(this, MainActivity::class.java)
         startActivity(intent)
         finish()
-    }
-
-    private fun updatePageDone(productName: String, pageName: String, item: MenuItem) {
-        var finishedPage: Boolean = false
-        val pageRef = database.child(productName).child(pageName)
-        val valueEventListener = object : ValueEventListener {
-            override fun onDataChange(dataSnapshot: DataSnapshot) {
-                for (ds in dataSnapshot.children) {
-                    val dbUsername = ds.key
-                    if (dbUsername == preferences.username) {
-                        finishedPage = true
-                    }
-                }
-                if (!finishedPage) {
-                    // Page completed
-                    database.child(productName).child(pageName).child(preferences.username)
-                        .setValue(true)
-                    item.setIcon(R.drawable.ic_baseline_check_24)
-                }
-            }
-
-            override fun onCancelled(databaseError: DatabaseError) {
-                // Here goes error message
-            }
-        }
-        pageRef.addListenerForSingleValueEvent(valueEventListener)
-    }
-
-    private fun checkDonePages(productName: String, pageName: String, item: MenuItem) {
-        val pageRef = database.child(productName).child(pageName)
-        val valueEventListener = object : ValueEventListener {
-            override fun onDataChange(dataSnapshot: DataSnapshot) {
-                for (ds in dataSnapshot.children) {
-                    if (ds.key == preferences.username && ds.value == true) {
-                        item.setIcon(R.drawable.ic_baseline_check_24)
-                    }
-                }
-            }
-            override fun onCancelled(databaseError: DatabaseError) {
-                // Here goes error message
-            }
-        }
-        pageRef.addListenerForSingleValueEvent(valueEventListener)
     }
 
     private fun replaceFragment(fragmentId: Int,fragment: Fragment) {
