+++ CircuitMessing/app/src/main/java/com/example/circuitmessing/products/makerbuino/MakerbuinoMeetTheToolsFragment.kt
@@ -1,4 +1,4 @@
-package com.example.circuitmessing.products.ringo
+package com.example.circuitmessing.products.makerbuino
 
 import androidx.lifecycle.ViewModelProvider
 import android.os.Bundle
@@ -8,24 +8,24 @@ import android.view.View
 import android.view.ViewGroup
 import com.example.circuitmessing.R
 
-class Ringo_introduction : Fragment() {
+class MakerbuinoMeetTheToolsFragment : Fragment() {
 
     companion object {
-        fun newInstance() = Ringo_introduction()
+        fun newInstance() = MakerbuinoMeetTheToolsFragment()
     }
 
-    private lateinit var viewModel: RingoIntroductionViewModel
+    private lateinit var viewModel: MakerbuinoMeetTheToolsViewModel
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.ringo_introduction_fragment, container, false)
+        return inflater.inflate(R.layout.makerbuino_meet_the_tools_fragment, container, false)
     }
 
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
-        viewModel = ViewModelProvider(this).get(RingoIntroductionViewModel::class.java)
+        viewModel = ViewModelProvider(this).get(MakerbuinoMeetTheToolsViewModel::class.java)
         // TODO: Use the ViewModel
     }
