+++ CircuitMessing/app/src/main/java/com/example/circuitmessing/products/makerbuino/MakerbuinoTimeToGetMakinFragment.kt
@@ -8,24 +8,24 @@ import android.view.View
 import android.view.ViewGroup
 import com.example.circuitmessing.R
 
-class Makerbuino_introduction : Fragment() {
+class MakerbuinoTimeToGetMakinFragment : Fragment() {
 
     companion object {
-        fun newInstance() = Makerbuino_introduction()
+        fun newInstance() = MakerbuinoTimeToGetMakinFragment()
     }
 
-    private lateinit var viewModel: MakerbuinoIntroductionViewModel
+    private lateinit var viewModel: MakerbuinoTimeToGetMakinViewModel
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.makerbuino_introduction_fragment, container, false)
+        return inflater.inflate(R.layout.makerbuino_time_to_get_makin_fragment, container, false)
     }
 
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
-        viewModel = ViewModelProvider(this).get(MakerbuinoIntroductionViewModel::class.java)
+        viewModel = ViewModelProvider(this).get(MakerbuinoTimeToGetMakinViewModel::class.java)
         // TODO: Use the ViewModel
     }
