+++ CircuitMessing/app/src/main/java/com/example/circuitmessing/products/nibble/NibbleTimeToGetMakinFragment.kt
@@ -1,4 +1,4 @@
-package com.example.circuitmessing.products.makerbuino
+package com.example.circuitmessing.products.nibble
 
 import androidx.lifecycle.ViewModelProvider
 import android.os.Bundle
@@ -8,24 +8,24 @@ import android.view.View
 import android.view.ViewGroup
 import com.example.circuitmessing.R
 
-class Makerbuino_time_to_get_makin : Fragment() {
+class NibbleTimeToGetMakinFragment : Fragment() {
 
     companion object {
-        fun newInstance() = Makerbuino_time_to_get_makin()
+        fun newInstance() = NibbleTimeToGetMakinFragment()
     }
 
-    private lateinit var viewModel: MakerbuinoTimeToGetMakinViewModel
+    private lateinit var viewModel: NibbleTimeToGetMakinViewModel
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.makerbuino_time_to_get_makin_fragment, container, false)
+        return inflater.inflate(R.layout.nibble_time_to_get_makin_fragment, container, false)
     }
 
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
-        viewModel = ViewModelProvider(this).get(MakerbuinoTimeToGetMakinViewModel::class.java)
+        viewModel = ViewModelProvider(this).get(NibbleTimeToGetMakinViewModel::class.java)
         // TODO: Use the ViewModel
     }
