+++ CircuitMessing/app/src/main/java/com/example/circuitmessing/products/ringo/RingoMeetTheToolsFragment.kt
@@ -1,4 +1,4 @@
-package com.example.circuitmessing.products.nibble
+package com.example.circuitmessing.products.ringo
 
 import androidx.lifecycle.ViewModelProvider
 import android.os.Bundle
@@ -8,24 +8,24 @@ import android.view.View
 import android.view.ViewGroup
 import com.example.circuitmessing.R
 
-class Nibble_meet_the_tools : Fragment() {
+class RingoMeetTheToolsFragment : Fragment() {
 
     companion object {
-        fun newInstance() = Nibble_meet_the_tools()
+        fun newInstance() = RingoMeetTheToolsFragment()
     }
 
-    private lateinit var viewModel: NibbleMeetTheToolsViewModel
+    private lateinit var viewModel: RingoMeetTheToolsViewModel
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.nibble_meet_the_tools_fragment, container, false)
+        return inflater.inflate(R.layout.ringo_meet_the_tools_fragment, container, false)
     }
 
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
-        viewModel = ViewModelProvider(this).get(NibbleMeetTheToolsViewModel::class.java)
+        viewModel = ViewModelProvider(this).get(RingoMeetTheToolsViewModel::class.java)
         // TODO: Use the ViewModel
     }
