+++ CircuitMessing/app/src/main/java/com/example/circuitmessing/products/ringo/RingoTimeToGetMakinFragmentFragment.kt
@@ -1,4 +1,4 @@
-package com.example.circuitmessing.products.nibble
+package com.example.circuitmessing.products.ringo
 
 import androidx.lifecycle.ViewModelProvider
 import android.os.Bundle
@@ -8,24 +8,28 @@ import android.view.View
 import android.view.ViewGroup
 import com.example.circuitmessing.R
 
-class Nibble_introduction_fragment : Fragment() {
+class RingoTimeToGetMakinFragmentFragment : Fragment() {
 
     companion object {
-        fun newInstance() = Nibble_introduction_fragment()
+        fun newInstance() = RingoTimeToGetMakinFragmentFragment()
     }
 
-    private lateinit var viewModel: NibbleIntroductionFragmentViewModel
+    private lateinit var viewModel: RingoTimeToGetMakinFragmentViewModel
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.nibble_introduction_fragment, container, false)
+        return inflater.inflate(
+            R.layout.ringo_time_to_get_makin_fragment,
+            container,
+            false
+        )
     }
 
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
-        viewModel = ViewModelProvider(this).get(NibbleIntroductionFragmentViewModel::class.java)
+        viewModel = ViewModelProvider(this).get(RingoTimeToGetMakinFragmentViewModel::class.java)
         // TODO: Use the ViewModel
     }
