+++ CircuitMessing/app/src/main/java/com/example/circuitmessing/ui/home/HomeFragment.kt
@@ -7,7 +7,6 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProviders
-import com.example.circuitmessing.*
 import com.example.circuitmessing.databinding.FragmentHomeBinding
 import com.example.circuitmessing.products.makerbuino.MakerbuinoActivity
 import com.example.circuitmessing.products.nibble.ActivityNibble
