+++ CircuitMessing/app/src/main/res/layout/fragment_login_fragment.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.auth.fragment_login">
+    tools:context=".ui.auth.FragmentLogin">
 
     <androidx.constraintlayout.widget.ConstraintLayout android:layout_width="match_parent" android:layout_height="350dp">
