+++ CircuitMessing/app/src/main/res/navigation/mobile_navigation.xml
@@ -25,12 +25,12 @@
 
     <fragment
         android:id="@+id/fragment_register"
-        android:name="com.example.circuitmessing.ui.auth.fragment_register"
+        android:name="com.example.circuitmessing.ui.auth.FragmentRegister"
         android:label="fragment_register_fragment"
         tools:layout="@layout/fragment_register_fragment" />
     <fragment
         android:id="@+id/fragment_login2"
-        android:name="com.example.circuitmessing.ui.auth.fragment_login"
+        android:name="com.example.circuitmessing.ui.auth.FragmentLogin"
         android:label="fragment_login_fragment"
         tools:layout="@layout/fragment_login_fragment" />
